agents/sot-agent/requirements.txt
snowflake-connector-python>=3.6.0
pyyaml>=6.0.1
openai>=1.40.0
anthropic>=0.36.0
tiktoken>=0.7.0
